pyproject.toml
[tool.poetry]
name = "repromatch"
version = "0.1.0"
description = ""
authors = ["Remi Rampin <remirampin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
jinja2 = "^2.11.2"
PyYAML = "^5.3.1"
markdown = "^3.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
